// packages/maskbook/src/plugins/RedPacket/types.ts
export enum RedPacketStatus {
    claimed = 'claimed',
    expired = 'expired',
    empty = 'empty',
    refunded = 'refunded',
}

export interface RedPacketAvailability {
    token_address: string
    balance: string
    total: string
    claimed: string
    expired: boolean
    claimed_amount?: string // V2
    ifclaimed?: boolean // V1
}

interface RedPacketBasic {
    contract_address: string
    rpid: string
    password: string
    shares: number
    is_random: boolean
    total: string
    creation_time: number
    duration: number
}

export interface RedPacketJSONPayload extends RedPacketBasic {
    sender: {
        address: string
        name: string
        message: string
    }
    txid?: string
    contract_version: number
    network?: string
    token_type: EthereumTokenType.Native | EthereumTokenType.ERC20
    token?: Pick<ERC20TokenRecord, 'address' | 'name' | 'decimals' | 'symbol'>
}
// ...
interface NFTRedPacketCreator extends Omit<RedPacketCreator, 'token' | 'is_random' | 'total_remaining'> {
    token: ERC721TokenDetailed
}

export interface RedPacketSubgraphInMask extends RedPacketBasic {
    message: string
    name: string
    txid: string
    total_remaining: string
    last_updated_time: number
    chain_id: number
    token: NativeTokenDetailed | ERC20TokenDetailed
    creator: RedPacketCreator
    claimers: {
        name: string
        address: string
    }[]
}

export interface NftRedPacketSubgraphInMask extends Omit<RedPacketSubgraphInMask, 'is_random' | 'token' | 'creator'> {
    token: ERC721TokenDetailed
    creator: NFTRedPacketCreator
    address: string
    token_ids: string[]
    claimers: {
        name: string
        address: string
    }[]
}

export interface RedPacketSubgraphOutMask extends Omit<RedPacketSubgraphInMask, 'token'> {
    token: TokenOutMask
}

export interface NftRedPacketSubgraphOutMask extends Omit<NftRedPacketSubgraphInMask, 'token'> {
    token: NftTokenOutMask
}

export interface RedPacketHistory extends RedPacketSubgraphInMask {
    payload: RedPacketJSONPayload
    contract_version: number
}
export interface NftRedPacketHistory extends NftRedPacketSubgraphInMask {
    payload: RedPacketJSONPayload
}